requirements.txt
chromadb
gradio
joblib
mediapipe
numpy
phi
Pillow
python-dotenv
Requests
tensorflow
opencv-contrib-python
opencv-python
phidata
python-dotenv
duckduckgo_search
openai
SpeechRecognition
gunicorn
